add move single group to page that sends object through params
