--- src/app/rsvp/group/[id]/page.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-export default async function Page({ params }) {
-    console.log('THIS IS PARAMS')
-    console.log(params)
-    
-        return (
-            <>
-                <div>
-                    SINGLE RSVP GROUP
-                </div>
-            </>
-        ) 
-}
